// crtr_angular/src/app/shared/models/Endereco.model.ts
export class Endereco {
  idEndereco: number;
  conteudo: any;
  idUsuario: number;
  flagAtivo: boolean = true;
  cep: string;
  logradouro: string;
  numero: string;
  complemento: string;
  bairro: string;
  localidade: string;
  uf: string;
  ibge: string;
  gia: string;
  idPessoa: number;
  enderecoUsuarioLogado: boolean;
  novoEndereco: boolean;
}